+++ src/inst/include/y2pm/InstTarget.h
@@ -213,10 +213,10 @@ class InstTarget: virtual public Rep, public InstData {
 	 * set callback function for reporting progress of package
 	 * installation
 	 *
-	 * @param func callback function, must accept double as argument
+	 * @param func callback function, must accept int as argument
 	 * @param data arbitrary data to pass when function is called
 	 * */
-	void setPackageInstallProgressCallback(void (*func)(double,void*), void* data);
+	void setPackageInstallProgressCallback(void (*func)(int,void*), void* data);
 
 	/**
 	 * @return destination root directory of target system
